fix block_size, clean up the expansion phase, do less work
Crypto Background:

HKDF generates the desired number of output bytes by running HMAC
multiple times (as a chained PRNG that includes an incrementing counter
for each block). HMAC itself is parameterized with a particular hash
function (so there's HMAC-SHA256, HMAC-SHA512, HMAC-MD5, etc). The
output size of HMAC for some particular hash function is the same as the
output of the underlying hash: 256 bits (32 bytes) for SHA256, 512
bits (64 bytes) for SHA512.

Independent of that, hash functions have an internal "block size" which
is not usually exposed. Hash functions are built around a low-level
"compression function" which takes a fixed-size input and modifies a
fixed-size state. The higher-level hash function takes an
arbitrary (unquantized) number of bytes as input, and knows how to feed
them to the compression function in block-sized chunks (adding padding
to the last block). The final output of the hash (in the common
Merkle–Damgård construction) is just the state vector after the last
block has been processed (which is why SHA2 and friends are vulnerable
to length-extension attacks).

The rust-crypto Digest trait exposes `output_bytes`, which is useful and
appropriate. Unfortunately it also exposes `block_size`, which is sort
of an attractive nuisance.

The `rust-hkdf` Fix:

The previous code was mistakenly using the hash's `block_size` to decide
how many times to call HMAC. It really should be using the HMAC
`output_bytes` value instead. This bug didn't affect correctness because
the old code worked by first concatenating the output of all the HMAC
calls, then finally trimming it to the desired length. The SHA256
`block_size` happens to be 64 bytes, which is larger than the SHA256 and
HMAC-SHA256 `output_bytes` (32). So the old code would generate twice as
much output as necessary, merge it all together, then truncate.

This new code reduces the number of HMAC calls to the minimal value. It
also removes one of the copies (and the merge+truncate phase) by
creating the return vector with the correct length ahead of time, and
copying HMAC outputs into the right offset as it goes. This is surely
faster, but I don't have benchmarks to prove it.

Finally there are some new tests to make sure I didn't screw up the
how-many-bytes-to-we-need checks, since now it matters more than it did
before.